refactor: type 宣言のファイルを切り分け
