+++ src/type/DisplayProductProps.ts
